pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adm"
dynamic = ["version"]
description = "EOSC Application Deployment Management API"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["EOSC", "Application", "Deployment", "Management", "API"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "fastapi==0.128.0",
    "pydantic==2.12.5",
    "pydantic[email]==2.12.5",
    "python-dotenv==1.2.1",
    "im-client>=1.8.2",
    "uvicorn[standard]==0.40.0",
    "requests-cache==1.2.1",
    "hvac==2.4.0",
    "cryptography==46.0.3",
]

[tool.setuptools.dynamic]
version = {attr = "adm.__version__"}

[project.optional-dependencies]
test = [
    "flask_testing==0.8.1",
    "coverage>=4.0.3",
    "nose>=1.3.7",
    "tox==3.20.1",
    "httpx==0.26.0",
    "pytest-mock",
    "pytest-cov",
]

[project.scripts]
adm = "adm.__main__:main"

[tool.setuptools]
packages = ["adm"]

[tool.setuptools.package-data]
adm = ["py.typed"]
